+++ src/connect/components/drive-settings-modal/add-public-drive-modal.tsx
@@ -2,7 +2,7 @@ import { DivProps, Modal } from '@/powerhouse';
 import { ComponentPropsWithoutRef } from 'react';
 import { twMerge } from 'tailwind-merge';
 import { Divider } from '..';
-import { AddPublicDriveForm } from '../drive-settings-form/add-public-drive-form';
+import { AddPublicDriveForm } from '../form/add-public-drive-form';
 
 type ModalProps = ComponentPropsWithoutRef<typeof Modal>;
 type FormProps = ComponentPropsWithoutRef<typeof AddPublicDriveForm>;
