+++ src/connect/components/index.ts
@@ -1,10 +1,10 @@
 export * from './breadcrumbs';
 export * from './divider';
-export * from './drive-settings-form';
 export * from './drive-settings-modal';
 export * from './drive-settings-select';
 export * from './drive-view';
 export * from './dropdown-menu';
+export * from './form';
 export * from './form-input';
 export * from './modal';
 export * from './search-bar';
